% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abbreviation}
\alias{abbreviation}
\title{The Dictionary for XJTLU Abbreviations.}
\format{
A data frame:
\describe{
  \item{short}{Abbreviation}
  \item{full}{Full name}
  ...
}
}
\usage{
abbreviation
}
\description{
A data table for looking up XJTLU abbreviations.
}
\keyword{datasets}
