% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_fullname}
\alias{get_fullname}
\title{Get the full name of an abbreviation}
\usage{
get_fullname(x = "")
}
\arguments{
\item{x}{short name}
}
\value{
full name
}
\description{
Get the full name of an abbreviation
}
\examples{
get_fullname(c("XJTLU", "HoD", "ENV"))
}
