% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_map}
\alias{get_map}
\title{Plot a campus map of XJTLU}
\usage{
get_map(english = FALSE)
}
\arguments{
\item{english}{logical. Display English on the map.}
}
\value{
a map
}
\description{
Plot a campus map of XJTLU
}
\examples{
get_map()
get_map(english = TRUE)
}
