% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{html_template}
\alias{html_template}
\alias{xjtlu_html_report}
\title{XJTLU html template}
\usage{
html_template(
  ...,
  template_name,
  toc = TRUE,
  code_folding = "hide",
  number_sections = TRUE
)

xjtlu_html_report(...)
}
\arguments{
\item{..., }{toc, code_folding, number_sections
Arguments passed to \code{rmarkdown::html_document()}.}

\item{template_name}{the template name.}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{code_folding}{Enable document readers to toggle the display of R code
chunks. Specify \code{"none"} to display all code chunks (assuming
they were knit with \code{echo = TRUE}). Specify \code{"hide"} to hide all R
code chunks by default (users can show hidden code chunks either
individually or document-wide). Specify \code{"show"} to show all R code
chunks by default.}

\item{number_sections}{\code{TRUE} to number section headings}
}
\value{
An R Markdown output format.
}
\description{
XJTLU html template
}
\details{
adapted from \url{https://github.com/holtzy/epuRate}.
}
