% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_bbb}
\alias{download_bbb}
\title{Download video recodings from BBB}
\usage{
download_bbb(url, destfolder = ".", if_merge = FALSE)
}
\arguments{
\item{url}{complete of the hyperlink of the webcam video}

\item{destfolder}{destination folder}

\item{if_merge}{whether merge.}
}
\value{
downloaded video files
}
\description{
Download video recodings from BBB
}
\examples{
download_bbb("https://bbbload.xjtlu.edu.cn/presentation/4c0413f756243e7b71221b81e36488f80d15996c-1599613283607/video/webcams.webm")
}
