% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_bbb}
\alias{download_bbb}
\title{Download video recodings from BBB}
\usage{
download_bbb(
  url = NA,
  camera_file = "webcams.webm",
  desktop_file = "desktop.webm",
  temp_file = "tmp.webm",
  merged_file = NA
)
}
\arguments{
\item{url}{The complete hyperlink of the webcam video. If NA, only merge the given videos.}

\item{camera_file}{The file name of the camera video.}

\item{desktop_file}{The file name of the desktop video.}

\item{temp_file}{The file name of the temporary file.}

\item{merged_file}{The file name of the merged file. If Na, do not merge.}
}
\value{
downloaded and/or merge video files
}
\description{
Download video recodings from BBB
}
\examples{
# download one
download_bbb("https://bbbload.xjtlu.edu.cn/presentation/4c0413f756243e7b71221b81e36488f80d15996c-1599613283607/video/webcams.webm")

# batch download and merge
urls <- c("https://bbbload.xjtlu.edu.cn/presentation/4c0413f756243e7b71221b81e36488f80d15996c-1599613283607/video/webcams.webm",
          "https://bbbload.xjtlu.edu.cn/presentation/4c0413f756243e7b71221b81e36488f80d15996c-1602031837942/video/webcams.webm",
          "https://bbbload.xjtlu.edu.cn/presentation/4c0413f756243e7b71221b81e36488f80d15996c-1600217953595/video/webcams.webm")
for (i in 1:3) {
  download_bbb(
    urls[i],
    camera_file = paste0("cam_chapter_", i, ".webm"),
    desktop_file = paste0("desk_chapter_", i, ".webm"),
    temp_file = paste0("tmp_chapter_", i, ".webm"),
    merged_file = paste0("bbb_chapter_", i, ".webm")
  )
}
}
